Add secondary keyboard matrix to the schematic.
The secondary matrix is driven by an ATTiny x61 which has the advantage of
having I2C pins separate from the SPI programming pins, this makes it easier
to program it without need for isolation from the ATMega32.

The single LED driver is used to light the LEDs on the secondary matrix as well
since this does not seem to have much influence in the final design requirements.