Implement breathing modes for leds.
Key (1,0) is used to cycle through leds.
Key (1,1) is used to trigger a scan of all the colors in 0x20 increments on the selected led.
Key (0,0) is used to trigger a keyboard half scan of all the colors in 0x20 increments.
Key (0,0) must first be pressed to trigger the open/short detection.